+++ .gitignore
@@ -26,4 +26,5 @@ chrome-user-data
 dist/
 .env
 postgresdata/
+server-dist/
 ssr
