+++ src/client/store/forum/index.ts
@@ -1,8 +1,8 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 
-import { ForumState } from 'store/forum/types';
-import { forumApi } from 'api/forum/forum-api';
-import {RootState} from 'store/types';
+import { ForumState } from "store/forum/types";
+import { forumApi } from "../../api/forum/forum-api";
+import {RootState} from '../types';
 
 const initialState: ForumState = {
   topic: null,
